Build a Mac OS X application bundle
Build a zip file containing a Mac OS X application bundle and place
alongside the jar file in the target directory.
On Mac OS X unzipping the file will result in an application that
can be started by double-clicking and which will get the correct
application name in the OS X system menu.