Don't use localization when writing events with SimpleGcWriter
PrintWriter.printf() uses the default locale of the system for formatting unless a locale is specified.
This affects e.g. numbers in locales that uses a comma instead of a period as decimal separator (and causes a failure in SimpleGcWriterTest when run on a system with Swedish locale).
The fix is to supply null as the locale since this will format without localization (according to the Javadoc for PrintWriter.printf()).