Command line graph generation random issue
The command line graph generation might fail due to a concurrency issue.
The SimpleChartRenderer is called by 'main' thread, it creates a
ModelChartImpl, initialize it and validates. And calls setScaleFactor
that calls repaint (suspicious). For some reason when paint is called on
the BufferedImage I get an half-drawn graph.
To be sure in any case, I rather do any swing call within the
EventDispatchThread