Fix a problem with quotes for language=fr
The quotes of the localized messages were not displayed and unit test "LocalisationHelperTest" crashed because the argument were not replaced.
- DataReaderFacade.java: use the LocalisationHelper.getString with arguments
instead of performing the format ourself.
- localStrings_fr.properties: replace each single quote with two single 
quotes as stated in the javadoc of MessageFormat class:
https://docs.oracle.com/javase/8/docs/api/java/text/MessageFormat.html#patterns
Otherwise the single quote won't appear in the message and the arguments are not parsed.