Allowing opening logfile series from "recent" files (follow up for #61)
When a series of logfiles was recently opened, it can now be reopened from the "recent" menu.
In order to support this, GCResourceGroup was refactored and now only stores the URL of a Resource instead of the whole resource (as it apparently was already before).
This allows to also store GcResourceSeries in the same way as regular GcResources.