Add Apache commons-lang3 to dependencies and change build to use maven-shade plugin
Apache commons-lang3 was added to the project dependencies. It contains many utilities that are useful and avoid having to rewrite them.
Among them is FastDateFormat, which is a thread-safe and fast alternative to SimpleDateFormat.

This dependency now has to be delivered together with GCViewer. The maven-shade plugin is now used to build an executable "fat-jar" that contains this and all other future dependencies.
The maven-jar plugin has been removed, as it is no longer required.